# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample todo-list server.",
        "title": "Todo-List API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/todos": {
            "get": {
                "description": "get the list of todos",
                "produces": [
                    "application/json"
                ],
                "summary": "Show all todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new todo",
                "parameters": [
                    {
                        "description": "Todo Content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AddTodoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Todo"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "get the single todo by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show a single todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TODO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Todo"
                        }
                    },
                    "404": {
                        "description": "no content"
                    }
                }
            },
            "put": {
                "description": "update a single todo by ID",
                "summary": "Update a todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Todo Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateTodoReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "404": {
                        "description": "no content"
                    }
                }
            },
            "delete": {
                "description": "delete a single todo by ID",
                "summary": "Delete a todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "404": {
                        "description": "no content"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AddTodoReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "main.Todo": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "done": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "main.UpdateTodoReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "done": {
                    "type": "boolean"
                }
            }
        }
    }
}